fem_parproj: small changes to prep for GPU impl. Main thing is switch set_rhs from using gkyl_superlu_brhs_from_triples to using gkyl_superlu_brhs_from_array. Fix definition of ghost in fem_poisson.
