+++ zero/gkyl_fem_parproj.h
@@ -29,11 +29,12 @@ typedef struct gkyl_fem_parproj gkyl_fem_parproj;
  * @param basis Basis functions of the DG field.
  * @param isparperiodic boolean indicating if parallel direction is periodic.
  * @param ctx Context for function evaluation. Can be NULL.
+ * @param use_gpu boolean indicating whether to use the GPU.
  * @return New updater pointer.
  */
 gkyl_fem_parproj* gkyl_fem_parproj_new(
   const struct gkyl_rect_grid *grid, const struct gkyl_basis *basis,
-  const bool isparperiodic, void *ctx);
+  const bool isparperiodic, void *ctx, bool use_gpu);
 
 /**
  * Set the multiplicative weight.
